cryptovote-web/requirements.txt
#
# ############# pip3 Requirements File #############
#                   Cryptovote-Web
# Example usage: pip3 install -r requirements.txt
#
flask
flask_sqlalchemy
python-dotenv
